# docs/swagger.yaml
definitions:
  common.JSONResult:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
...
  pkg.Application:
    properties:
      abandoned:
        type: boolean
      candidate_id:
        description: manytoone
        type: string
      comments:
        description: onetomany
        items:
          $ref: '#/definitions/pkg.Comment'
        type: array
      created_at:
        type: string
      grade:
        description: pkg.Grade
        type: string
      group:
        allOf:
        - $ref: '#/definitions/pkg.Group'
        description: pkg.Group
      institute:
        type: string
      interview_allocations_group:
        type: string
      interview_allocations_team:
        type: string
      interview_selections:
        description: manytomany
        items:
          $ref: '#/definitions/pkg.Interview'
        type: array
      intro:
        type: string
      is_quick:
        type: boolean
      major:
        type: string
      rank:
        type: string
      recruitment_id:
        description: manytoone
        type: string
      referrer:
        type: string
      rejected:
        type: boolean
      resume:
        type: string
      step:
        allOf:
        - $ref: '#/definitions/pkg.Step'
        description: pkg.Step
      uid:
        type: string
      updated_at:
        type: string
      user_detail:
        allOf:
        - $ref: '#/definitions/pkg.UserDetail'
        description: get from sso
    type: object
  pkg.Comment:
    properties:
      application_id:
        description: manytoone
        type: string
      content:
        type: string
      created_at:
        type: string
      evaluation:
        $ref: '#/definitions/pkg.Evaluation'
      member_id:
        description: manytoone
        type: string
      uid:
        type: string
      updated_at:
        type: string
    type: object
...
  pkg.CreateCommentOpts:
    properties:
      application_id:
        type: string
      content:
        type: string
      evaluation:
        $ref: '#/definitions/pkg.Evaluation'
      member_id:
        type: string
    required:
    - application_id
    - content
    - evaluation
    type: object
  pkg.CreateRecOpts:
    properties:
      beginning:
        type: string
      deadline:
        type: string
      end:
        type: string
      name:
        type: string
    required:
    - beginning
    - deadline
    - end
    - name
    type: object
  pkg.Evaluation:
    enum:
    - 0
    - 1
    - 2
    type: integer
    x-enum-varnames:
    - Good
    - Normal
    - Bad
  pkg.Gender:
    enum:
    - 1
    - 2
    - 3
    type: integer
    x-enum-varnames:
    - Male
    - Female
    - Oth
  pkg.Group:
    enum:
    - web
    - lab
    - ai
    - game
    - android
    - ios
    - design
    - pm
    - unique
    type: string
    x-enum-comments:
      Unique: for team interview
    x-enum-varnames:
    - Web
    - Lab
    - Ai
    - Game
    - Android
    - Ios
    - Design
    - Pm
    - Unique
  pkg.Interview:
    properties:
      applications:
        description: manytomany
        items:
          $ref: '#/definitions/pkg.Application'
        type: array
      created_at:
        type: string
      date:
        type: string
      end:
        type: string
      name:
        $ref: '#/definitions/pkg.Group'
      period:
        $ref: '#/definitions/pkg.Period'
      recruitment_id:
        description: manytoone
        type: string
      select_number:
        type: integer
      slot_number:
        type: integer
      start:
        type: string
      uid:
        type: string
      updated_at:
        type: string
    type: object
  pkg.Period:
    enum:
    - morning
    - afternoon
    - evening
    type: string
    x-enum-varnames:
    - Morning
    - Afternoon
    - Evening
  pkg.Recruitment:
    properties:
      applications:
        description: 一个hr->简历 ;级联删除
        items:
          $ref: '#/definitions/pkg.Application'
        type: array
      beginning:
        type: string
      created_at:
        type: string
      deadline:
        type: string
      end:
        type: string
      interviews:
        description: 一个hr->面试 ;级联删除
        items:
          $ref: '#/definitions/pkg.Interview'
        type: array
      name:
        type: string
      statistics:
        type: string
      uid:
        type: string
      updated_at:
        type: string
    type: object
...
  pkg.Step:
    enum:
    - SignUp
    - WrittenTest
    - GroupTimeSelection
    - GroupInterview
    - OnlineGroupInterview
    - StressTest
    - TeamTimeSelection
    - TeamInterview
    - OnlineTeamInterview
    - Pass
    type: string
    x-enum-comments:
      GroupInterview: 组面
      GroupTimeSelection: 组面时间选择
      OnlineGroupInterview: 在线组面
      OnlineTeamInterview: 在线群面
      Pass: 通过
      SignUp: 报名
      StressTest: 熬测
      TeamInterview: 群面
      TeamTimeSelection: 面试时间选择
      WrittenTest: 笔试
    x-enum-varnames:
    - SignUp
    - WrittenTest
    - GroupTimeSelection
    - GroupInterview
    - OnlineGroupInterview
    - StressTest
    - TeamTimeSelection
    - TeamInterview
    - OnlineTeamInterview
    - Pass
...
  pkg.UpdateInterviewOpts:
    properties:
      date:
        type: string
      end:
        type: string
      period:
        $ref: '#/definitions/pkg.Period'
      slot_number:
        type: integer
      start:
        type: string
      uid:
        type: string
    required:
    - date
    - end
    - period
    - slot_number
    - start
    type: object
  pkg.UpdateRecOpts:
    properties:
      beginning:
        type: string
      deadline:
        type: string
      end:
        type: string
      name:
        type: string
      rid:
        type: string
    type: object
  pkg.UserDetail:
    properties:
      avatar_url:
        type: string
      email:
        type: string
      gender:
        $ref: '#/definitions/pkg.Gender'
      groups:
        items:
          type: string
        type: array
      join_time:
        type: string
      lark_union_id:
        type: string
      name:
        type: string
      password:
        type: string
      phone:
        type: string
      uid:
        type: string
    type: object
...
host: https://dev.back.recruitment2023.hustunique.com/
info:
  contact:
    email: wwbstar07@gmail.com
  description: This is API doc of UniqueStudio Recruitment. For more API information,
    please see https://app.apifox.com/project/2985744
  title: UniqueStudio Recruitment API
  version: "0.1"
paths:
...
  /comments:
    post:
      consumes:
      - application/json
      description: create comment for applications, only can be created by member.
      operationId: create_comment.
      parameters:
      - description: create comment opts
        in: body
        name: pkg.CreateCommentOpts
        required: true
        schema:
          $ref: '#/definitions/pkg.CreateCommentOpts'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.JSONResult'
            - properties:
                data:
                  $ref: '#/definitions/pkg.Comment'
              type: object
        "400":
          description: code is not 0 and msg not empty
          schema:
            $ref: '#/definitions/common.JSONResult'
      summary: create comment for application
      tags:
      - comment
  /comments/{cid}:
    delete:
      consumes:
      - application/json
      description: delete comment of application, only can be deleted by comment's
        owner.
      operationId: delete_comment.
      parameters:
      - description: comment uid
        in: path
        name: cid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.JSONResult'
        "400":
          description: code is not 0 and msg not empty
          schema:
            $ref: '#/definitions/common.JSONResult'
      summary: delete comment of application
      tags:
      - comment
  /recruitments:
    get:
      consumes:
      - application/json
      description: get pending(the most recent) recruitment, member can only get the
        recruitment's detail(include application, interviews) after join in.
      operationId: get_pending_recruitment.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.JSONResult'
            - properties:
                data:
                  $ref: '#/definitions/pkg.Recruitment'
              type: object
        "400":
          description: code is not 0 and msg not empty
          schema:
            $ref: '#/definitions/common.JSONResult'
      summary: get pending recruitment.
      tags:
      - recruitment
    post:
      consumes:
      - application/json
      description: gcreate recruitment, only can be created by admin
      operationId: create_recruitment.
      parameters:
      - description: create recruitment opts
        in: body
        name: pkg.CreateRecOpts
        required: true
        schema:
          $ref: '#/definitions/pkg.CreateRecOpts'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.JSONResult'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/pkg.Recruitment'
                  type: array
              type: object
        "400":
          description: code is not 0 and msg not empty
          schema:
            $ref: '#/definitions/common.JSONResult'
      summary: create recruitment.
      tags:
      - recruitment
  /recruitments/{rid}:
    get:
      consumes:
      - application/json
      description: get recruitment, member can only get the recruitment's detail(include
        application, interviews) after join in.
      operationId: get_recruitment.
      parameters:
      - description: recruitment uid
        in: path
        name: rid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.JSONResult'
            - properties:
                data:
                  $ref: '#/definitions/pkg.Recruitment'
              type: object
        "400":
          description: code is not 0 and msg not empty
          schema:
            $ref: '#/definitions/common.JSONResult'
      summary: get recruitment.
      tags:
      - recruitment
    put:
      consumes:
      - application/json
      description: update recruitment, only can be updated by admin
      operationId: update_recruitment.
      parameters:
      - description: recruitment uid
        in: path
        name: rid
        required: true
        type: string
      - description: update recruitment opts
        in: body
        name: pkg.UpdateRecOpts
        required: true
        schema:
          $ref: '#/definitions/pkg.UpdateRecOpts'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.JSONResult'
        "400":
          description: code is not 0 and msg not empty
          schema:
            $ref: '#/definitions/common.JSONResult'
      summary: update recruitment.
      tags:
      - recruitment
  /recruitments/{rid}/interviews/{name}:
    get:
      consumes:
      - application/json
      description: get recruitment interviews, only can be got by member(will get
        interviews of groups or unique)
      operationId: get_recruitment_interviews.
      parameters:
      - description: recruitment id
        in: path
        name: rid
        required: true
        type: string
      - description: pkg.Group or unique
        enum:
        - web
        - lab
        - ai
        - game
        - android
        - ios
        - design
        - pm
        - unique
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.JSONResult'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/pkg.Interview'
                  type: array
              type: object
        "400":
          description: code is not 0 and msg not empty
          schema:
            $ref: '#/definitions/common.JSONResult'
      summary: get recruitment interviews.
      tags:
      - interviews
    put:
      consumes:
      - application/json
      description: get recruitment interviews, use PUt method to prevent resource
        are duplicated
      operationId: set_recruitment_interviews.
      parameters:
      - description: recruitment id
        in: path
        name: rid
        required: true
        type: string
      - description: pkg.Group or unique
        enum:
        - web
        - lab
        - ai
        - game
        - android
        - ios
        - design
        - pm
        - unique
        in: path
        name: name
        required: true
        type: string
      - description: update interview info
        in: body
        name: '[]pkg.UpdateInterviewOpts'
        required: true
        schema:
          items:
            $ref: '#/definitions/pkg.UpdateInterviewOpts'
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.JSONResult'
        "400":
          description: code is not 0 and msg not empty
          schema:
            $ref: '#/definitions/common.JSONResult'
      summary: set recruitment interviews.
      tags:
      - interviews
...
swagger: "2.0"